graphs on monitor page now makes sense when reloading page during round + refactor
