+++ market/templates/market/monitor.html
@@ -127,6 +127,10 @@
     </div>
 </div>
 
+<div class="d-flex justify-content-center mt-5">
+    <a class="btn btn-secondary" href="{% url 'market:monitor' market.market_id %}">Update graphs</a>
+</div>
+
 <h4>Graphs</h4>
 <div id="accordion">
     <div class="card">
@@ -137,9 +141,9 @@
                 </button>
             </h5>
         </div>
-        <div id="collapse_balance" class="collapse" aria-labelledby="heading_balance" data-parent="#accordion">
+        <div id="collapse_balance" class="collapse show" aria-labelledby="heading_balance">
             <div class="card-body">  
-                <div class="mb-5">
+                <div>
                     <canvas id="balanceCanvas"></canvas>
                 </div>
             </div>
@@ -153,7 +157,7 @@
                 </button>
             </h5>
         </div>
-        <div id="collapse_price" class="collapse" aria-labelledby="heading_price" data-parent="#accordion">
+        <div id="collapse_price" class="collapse show" aria-labelledby="heading_price">
             <div class="card-body">
                 <div class="mb-5">
                     <canvas id="priceCanvas"></canvas>
@@ -169,9 +173,9 @@
                 </button>
             </h5>
         </div>
-        <div id="collapse_amount" class="collapse" aria-labelledby="heading_amount" data-parent="#accordion">
+        <div id="collapse_amount" class="collapse show" aria-labelledby="heading_amount">
             <div class="card-body">
-                <div class="mb-5">
+                <div>
                     <canvas id="amountCanvas"></canvas>
                 </div>
             </div>
@@ -180,6 +184,7 @@
 </div>
 <br><br>
 
+{% comment %} 
 <h4>Tabular Data</h4>
 <div id="accordion">
     {% for field in show_stats_fields %}
@@ -231,19 +236,20 @@
                                 {% endfor %}
                             </tbody>
                         </table>
-                    </div>
-                {% else %}
+                        </div>
+                        {% else %}
                     <p class="text-muted">No data to show yet.</p>
                 {% endif %}
-            </div>
-        </div>
-    </div>
-    {% endfor %}
-</div> 
-<br><br>
+                </div>
+                </div>
+                </div>
+                {% endfor %}
+                </div>
+                <br><br> {% endcomment %}
 
-{% comment %} <div class="d-flex justify-content-center">
-    <button type="button" class="btn btn-primary mb-5" id="next_round_btn" onclick="alert('Feature not implemented yet, sorry')">
+                {% comment %} <div class="d-flex justify-content-center">
+                    <button type="button" class="btn btn-primary mb-5" id="next_round_btn"
+                        onclick="alert('Feature not implemented yet, sorry')">
         Download Data
     </button>
 </div> {% endcomment %}
@@ -251,7 +257,6 @@
 {% endblock %}
 
 {% block javascript %}
-
 <script>
     function copy_join_link() {
         // From w3schools
@@ -265,15 +270,34 @@
 <!-- import Chart.js library-->
 <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 
-<!-- Settings for the three charts -->
+<!-- Chart settings -->
 <script>
  
     // Global settings for charts on this page
     Chart.defaults.plugins.title.display = true;
     Chart.defaults.plugins.title.padding = {top:5, bottom:10};
-    Chart.defaults.scales.linear.beginAtZero = true; // y axes is forced to include 0    
-  
+    Chart.defaults.scales.linear.beginAtZero = true; // y axes is forced to include 0
     
+    function responsive_aspectRatio(){
+        // Set x-axis height relative to y-axis
+        if (window.innerWidth < 600){
+            return 1.3 
+        }
+        else{ 
+            return 2 
+        } 
+    } 
+        
+    function reponsive_display_y_axis_title(){ 
+        // Only show y-axis legend on big enough screens 
+        if (window.innerWidth < 800) { 
+            return false 
+            } 
+            else{
+                 return true
+        }
+    }
+
     function format_amount_labels(value, index, values){
         // Default format of 5000 on axis is 5,000. 
         // Here we change this to 5000.00
@@ -281,8 +305,12 @@
     }
 
     // Balance chart 
-    balance_labels = JSON.parse("{{ round_labels_json }}"), // labels on x-axis
-    balance_labels.push("Final")
+    balance_labels = JSON.parse("{{ round_labels_json }}") // labels on x-axis
+    
+    {% if not market.endless %}
+        balance_labels.push("Final")
+    {% endif %}
+
     var balanceChart = new Chart(document.getElementById('balanceCanvas'), {
         type: 'line',
         data: {
@@ -290,11 +318,12 @@
             datasets: {{ balanceDataSet|safe }}
         },
         options: {
+            aspectRatio: responsive_aspectRatio(),
             scales: {
                 y: {
                     title:{ 
                         text: 'Balance (kr.)',
-                        display: 'true'
+                        display: reponsive_display_y_axis_title()
                     },
                     ticks: {
                         callback: format_amount_labels
@@ -324,17 +353,17 @@
             datasets: {{ priceDataSet|safe }}
         },
         options: {
+            aspectRatio: responsive_aspectRatio(),
             scales: {
                 y: {
                     title:{ 
                         text: 'Unit price (kr.)',
-                        display: 'true'
+                        display: reponsive_display_y_axis_title()
                     },
                     ticks: {
                         callback: format_amount_labels
                     },
                     suggestedMax: 2*parseInt("{{ market.max_cost }}"),
-
                 },
                 x: {
                     title:{
@@ -360,11 +389,12 @@
             datasets: {{ amountDataSet|safe }}
         },
         options:{
+            aspectRatio: responsive_aspectRatio(),
             scales: {
                 y: {
                     title:{ 
-                        text: 'Units produced',
-                        display: 'true'
+                        text: 'Number of units',
+                        display: reponsive_display_y_axis_title()
                     },
                     suggestedMax: 100,
